fix: Authorization by claims userId

+++ code/server/SocialMedia/SM.WebApi/Endpoints/CommentEndpoints.cs
@@ -5,6 +5,7 @@ using SM.Application.Comments.DeleteComment;
 using SM.Application.Comments.GetCommentById;
 using SM.Application.Comments.GetComments;
 using SM.Application.Shared.Models;
+using SM.WebApi.Extensions;
 
 namespace SM.WebApi.Endpoints;
 
@@ -22,14 +23,24 @@ public static class CommentEndpoints
         return group;
     }
 
-    public static async Task<IResult> CreateComment([FromBody] CreateCommentCommand command, ISender sender)
+    public static async Task<IResult> CreateComment([FromBody] CreateCommentCommand command, ISender sender, HttpContext httpContext)
     {
+        Guid userId = httpContext.GetLoggedInUserId();
+        if (userId == Guid.Empty)
+            return TypedResults.Unauthorized();
+
+        command = command with { AuthorId = userId };
+
         var createdComment = await sender.Send(command);
         return TypedResults.Ok(createdComment);
     }
 
-    public static async Task<IResult> DeleteComment(Guid id, ISender sender)
+    public static async Task<IResult> DeleteComment(Guid id, ISender sender, HttpContext httpContext)
     {
+        Guid userId = httpContext.User.GetUserId();
+        if (userId == Guid.Empty)
+            return TypedResults.Unauthorized();
+
         var deletedComment = await sender.Send(new DeleteCommentCommand(id));
         return TypedResults.Ok(deletedComment);
     }
